preview/claude-company-benefits-scraper-dk7TZ/server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiomysql==0.2.0
cryptography>=42.0.0
pydantic[email]==2.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
